pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "pylint-pytest>=1.1.8",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.6",
]
